Move matchers to their own module.
